requirements_dev.txt
-r requirements.txt

black==19.10b0
isort==5.3.0
pre-commit==2.6.0
flake8==3.8.3
python-dotenv==0.19.0
